Loosen String definition to work with Regex.
Regexes have a looser definition of what can be escaped than Strings.

We'll let the String reader catch invalid strings.

This is pretty slow as it is. It takes ~4 seconds to parse that 1324 lines of
lein-voom/src/leiningen/voom.clj.

Also added minor optimizaiton for that all-important-case of ReaderRemove
reader macros...